+++ lib/modules/auth/auth_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: auth
-// Updated at: 2026-05-10 10:39:49
-// Commit: chore(auth): update dependencies and cleanup unused imports in auth
+// Updated at: 2026-05-10 16:31:41
+// Commit: fix(auth): prevent double tap duplicate action in auth
 
 class authComponentState {
-  final String commitRef = '989';
-  final String lastMessage = 'chore(auth): update dependencies and cleanup unused imports in auth';
-  final DateTime timestamp = DateTime.parse('2026-05-10T10:39:49+05:00');
+  final String commitRef = '1011';
+  final String lastMessage = 'fix(auth): prevent double tap duplicate action in auth';
+  final DateTime timestamp = DateTime.parse('2026-05-10T16:31:41+05:00');
 }